03_Naive_RAG/requirements.txt
streamlit==1.28.0
google-generativeai==0.8.0
chromadb==0.4.22
PyPDF2==3.0.1
langchain==0.3.7
